// Portal.Portal.Web/src/app/shared/infrastructure/PortalHttpClient/model/resourceResponse.ts
/**
 * Demo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuthorizationModuleResourceModel } from './authorizationModuleResourceModel';
import { EmploymentModuleResourceModel } from './employmentModuleResourceModel';
import { ErrorResourceModel } from './errorResourceModel';
import { LayoutResourcesModel } from './layoutResourcesModel';
import { MonthResourcesModel } from './monthResourcesModel';
import { NonComplianceResourcesModel } from './nonComplianceResourcesModel';
import { ParametresResourceModel } from './parametresResourceModel';
import { RolesModuleResourcesModel } from './rolesModuleResourcesModel';
import { TimeOffRequestResourcesModel } from './timeOffRequestResourcesModel';

export interface ResourceResponse { 
    errorResources?: ErrorResourceModel;
    employmentModuleResources?: EmploymentModuleResourceModel;
    authorizationModuleResources?: AuthorizationModuleResourceModel;
    timeOffRequestResources?: TimeOffRequestResourcesModel;
    layoutResources?: LayoutResourcesModel;
    monthResources?: MonthResourcesModel;
    nonComplianceResources?: NonComplianceResourcesModel;
    parametresResources?: ParametresResourceModel;
    rolesResources?: RolesModuleResourcesModel;
}